fix: Authorization Bypass Vulnerability
Not using `pam_acct_mgmt` after `pam_authenticate` to check the validity
of a login can lead to an authorization bypass

In this case, only a call to `pam.Authenticate` is used to login a user.
This implies that a user with expired credentials can still login.
The bug can be verified easily by creating a new user account,
expiring it with <code>chage -E0 `username` </code> and then trying to
log in with the expired credentials.

Log: Fix Authorization Bypass